+++ tests/api_contract/fail/constants_module_is_internal.rs
@@ -1,3 +1,0 @@
-fn main() {
-    let _ = grafton_visca::constants::CameraVariant::PtzOpticsG2;
-}