+++ tests/api_contract/fail/response_types_module_is_internal.rs
@@ -1,3 +1,6 @@
 fn main() {
     let _ = core::any::TypeId::of::<grafton_visca::command::response::types::Response>();
 }
+
+//~ E0603
+//~ "module `response` is private"