+++ tests/api_contract/fail_no_test_utils/testing_module_requires_test_utils.rs
@@ -1,3 +1,6 @@
 fn main() {
     let _ = core::any::TypeId::of::<grafton_visca::testing::testkit::DeterministicExecutor>();
 }
+
+//~ E0433
+//~ "cannot find `testing` in `grafton_visca`"